requirements.txt
robotframework
robotframework-seleniumlibrary
selenium
webdriver-manager
robotframework-faker
httpx
